// src/types/profile/profiletype.ts
export type ProfileData = {
  countryCode: string;
  deleted: boolean;
  email: string;
  firstname: string;
  gender: number | string;
  id: string;
  isLoggedIn: boolean;
  isVerified: boolean;
  lastname: string;
  mobileNo: string;
  role_id: number;
  status: string;
  user_type: number;
  username: string;
};